+++ web/app/themes/helsinki-testbed/resources/assets/scripts/editor.js
@@ -43,6 +43,20 @@ const withDefaultBackgroundColor = createHigherOrderComponent(BlockListBlock => 
   };
 }, 'withDefaultBackgroundColor' );
 /**
+ * Sets outline as default style for buttons.
+ */
+const withDefaultButtonStyle = createHigherOrderComponent(BlockListBlock => {
+  return (props) => {
+    if (
+      props.name === 'core/button' &&
+      props.attributes.style !== 'outline'
+    ) {
+      props.attributes.style = 'outline';
+    }
+    return <BlockListBlock { ...props } />;
+  };
+}, 'withDefaultButtonStyle' );
+/**
  * Sets default color for buttons with outline style.
  */
 const withDefaultColor = createHigherOrderComponent(BlockListBlock => {
@@ -72,12 +86,84 @@ const withRemoveBorderRadius = createHigherOrderComponent(BlockListBlock => {
   };
 }, 'withRemoveBorderRadius' );
 
+var excludeBlockTypes = [
+  'core/archives',
+  'core/calendar',
+  'core/categories',
+  'core/latest-posts',
+  'core/navigation',
+  'core/nextpage',
+  'core/latest-comments',
+  'core/more',
+  'core/rss',
+  'core/search',
+  'core/social-links',
+  'core/table',
+  'core/tag-cloud',
+  'core-embed/amazon-kindle',
+  'core-embed/animoto',
+  'core-embed/cloudup',
+  'core-embed/collegehumor',
+  'core-embed/crowdsignal',
+  'core-embed/dailymotion',
+  'core-embed/facebook',
+  'core-embed/flickr',
+  'core-embed/hulu',
+  'core-embed/imgur',
+  'core-embed/instagram',
+  'core-embed/issuu',
+  'core-embed/kickstarter',
+  'core-embed/meetup-com',
+  'core-embed/mixcloud',
+  'core-embed/polldaddy',
+  'core-embed/reddit',
+  'core-embed/reverbnation',
+  'core-embed/screencast',
+  'core-embed/scribd',
+  'core-embed/slideshare',
+  'core-embed/smugmug',
+  'core-embed/soundcloud',
+  'core-embed/speaker',
+  'core-embed/speaker-deck',
+  'core-embed/spotify',
+  'core-embed/ted',
+  'core-embed/tiktok',
+  'core-embed/tumblr',
+  'core-embed/twitter',
+  'core-embed/videopress',
+  'core-embed/vimeo',
+  'core-embed/wordpress',
+  'core-embed/wordpress-tv',
+  'core-embed/youtube',
+];
+wp.hooks.addFilter('blocks.registerBlockType', 'pw-examples/exclude-blocks', function(settings, name) {
+  if (excludeBlockTypes.indexOf(name) !== -1) {
+      return Object.assign({}, settings, {
+          supports: Object.assign({}, settings.supports, {inserter: false}),
+      });
+  }
+  return settings;
+});
+
+function getBackgroundShapeBlockList (BlockListBlock) {
+  return (props) => {
+    return (
+      <BlockListBlock {...props} />
+    )
+  }
+}
+addFilter(
+  'editor.BlockListBlock',
+  'hds/print-block-list',
+  createHigherOrderComponent(getBackgroundShapeBlockList)
+)
+
 domReady(() => {
   unregisterBlockStyle('core/button', 'fill');
-
   registerBlockStyle('core/button', {
     name: 'outline',
     label: 'Outline',
+    isDefault: true
   });
 
   registerBlockStyle('core/media-text', {
@@ -117,5 +203,9 @@ domReady(() => {
   addFilter(
     'editor.BlockListBlock', 'sage/with-remove-border-radius',
     withRemoveBorderRadius
+  );
+  addFilter(
+    'editor.BlockListBlock', 'sage/with-default-button-style',
+    withDefaultButtonStyle
   );
 });